Fix issue where -track_input collects all zeros in realtime collection.
This was a general issue that only impacted -track_input.  If the event
all keyword mask became zero due to disjoint masks from different
events, then it followed the initial/clear path and would accept the
next event mask as the all-mask.

Instead, we use any-mask==0 to indicate that no masks are set.