Fix for msBetweenConsumerPresents when consumer is a frame or higher behind the consumer
When a workload is consumer bound, the consumer can fall behind the
producer by several frames.  When this happens, PresentMon cannot
properly associate frames flowing through the system with the correct
Present() call as events aren't necessarily associated with the most
recent Present() call anymore.

This change fixes an issue when the consumer is close to a frame
multiple behind.  In this case work variation can cause the consumer to
vary between n and n-1 frames behind which was causing one Present()'s
data to be overwritten, and then the next to have no data.

This change does not fix the overall present assignment problem, so
latency measurements are wrong when the consumer is more than a frame
behind.