Allow MMIOFlipMPO_Info event to complete the GPU frame
It's possible for MMIOFlipMPO_Info to occur before the present's
QueuePacket_Stop (e.g., if the GPU rendering is already complete).

Before this change, such a MMIOFlipMPO_Info event would cause the
present to be completed before the tracked GPU work was assigned to it,
and any tracked GPU work would be accumulated onto the next frame
instead.  This resulted in a GPUActive metric value of 0 followed by a
present with GPUActive being 2x too large.

See discussion in internal issue #1.