Handle wait packets when tracking GPU work on HWS
When HWS is being used we use the packet queue instead of the dma queue
but before this commit wait packets were being ignored.

That could cause situations where PresentMon would incorrectly consider
a packet running when it was really blocked waiting for a fence, and
some portion of the wait time would be incorrectly attributed to
GPUActive.